+++ lib/td_dd_web/controllers/user_search_filter_controller.ex
@@ -2,6 +2,7 @@ defmodule TdDdWeb.UserSearchFilterController do
   use TdDdWeb, :controller
   use PhoenixSwagger
 
+  alias TdCache.Permissions
   alias TdDd.UserSearchFilters
   alias TdDd.UserSearchFilters.UserSearchFilter
   alias TdDdWeb.ErrorView
@@ -37,11 +38,20 @@ defmodule TdDdWeb.UserSearchFilterController do
   end
 
   def index_by_user(conn, params) do
-    %{user_id: user_id} = conn.assigns[:current_resource]
-    params_with_user = Map.put(params, "user_id", user_id)
+    %{user_id: user_id} = claims = conn.assigns[:current_resource]
+
+    global_domains = permitted_view_domains(claims)
+    params_with_user = params
+    |> Map.put("user_id", user_id)
+    |> Map.put("with_globals", global_domains)
 
     user_search_filters = UserSearchFilters.list_user_search_filters(params_with_user)
     render(conn, "index.json", user_search_filters: user_search_filters)
+  end
+
+  defp permitted_view_domains(%{role: "admin"}), do: :all
+  defp permitted_view_domains(%{jti: jti}) do
+    Permissions.permitted_domain_ids(jti, :view_data_structure)
   end
 
   swagger_path :create do
@@ -62,11 +72,12 @@ defmodule TdDdWeb.UserSearchFilterController do
   end
 
   def create(conn, %{"user_search_filter" => user_search_filter_params}) do
-    %{user_id: user_id} = conn.assigns[:current_resource]
+    %{user_id: user_id} = claims = conn.assigns[:current_resource]
 
     create_params = Map.put(user_search_filter_params, "user_id", user_id)
 
-    with {:ok, %UserSearchFilter{} = user_search_filter} <-
+    with :ok <- Bodyguard.permit(UserSearchFilters, :create, claims, create_params),
+      {:ok, %UserSearchFilter{} = user_search_filter} <-
            UserSearchFilters.create_user_search_filter(create_params) do
       conn
       |> put_status(:created)